Split out emissions in prevision for VHMM.
