+++ lib/hmmlearn/tests/test_categorical_hmm.py
@@ -140,13 +140,9 @@ class TestCategoricalHMM:
     @pytest.mark.parametrize("implementation", ["scaling", "log"])
     def test__check_and_set_categorical_n_features(self, implementation):
         h = self.new_hmm(implementation)
-        h._check_and_set_categorical_n_features(
-            np.array([[0, 0, 2, 1, 3, 1, 1]]))
-        h._check_and_set_categorical_n_features(
-            np.array([[0, 0, 1, 3, 1]], np.uint8))
+        h._check_and_set_n_features(np.array([[0, 0, 2, 1, 3, 1, 1]]))
+        h._check_and_set_n_features(np.array([[0, 0, 1, 3, 1]], np.uint8))
         with pytest.raises(ValueError):  # non-integral
-            h._check_and_set_categorical_n_features(
-                np.array([[0., 2., 1., 3.]]))
+            h._check_and_set_n_features(np.array([[0., 2., 1., 3.]]))
         with pytest.raises(ValueError):  # negative integers
-            h._check_and_set_categorical_n_features(
-                np.array([[0, -2, 1, 3, 1, 1]]))
+            h._check_and_set_n_features(np.array([[0, -2, 1, 3, 1, 1]]))
