+++ lib/hmmlearn/tests/test_multinomial_hmm.py
@@ -114,20 +114,20 @@ class TestMultinomialHMM:
             self, implementation):
         h = hmm.MultinomialHMM(
             n_components=2, n_trials=None, implementation=implementation)
-        h._check_and_set_multinomial_n_features_n_trials(
+        h._check_and_set_n_features(
             np.array([[0, 2, 3, 0], [1, 0, 2, 2]]))
         assert (h.n_trials == 5).all()
         with pytest.raises(ValueError):  # wrong dimensions
-            h._check_and_set_multinomial_n_features_n_trials(
+            h._check_and_set_n_features(
                 np.array([[0, 0, 2, 1, 3, 1, 1]]))
         with pytest.raises(ValueError):  # not added up to n_trials
-            h._check_and_set_multinomial_n_features_n_trials(
+            h._check_and_set_n_features(
                 np.array([[0, 0, 1, 1], [3, 1, 1, 0]]))
         with pytest.raises(ValueError):  # non-integral
-            h._check_and_set_multinomial_n_features_n_trials(
+            h._check_and_set_n_features(
                 np.array([[0., 2., 0., 3.], [0.0, 2.5, 2.5, 0.0]]))
         with pytest.raises(ValueError):  # negative integers
-            h._check_and_set_multinomial_n_features_n_trials(
+            h._check_and_set_n_features(
                 np.array([[0, -2, 1, 6], [5, 6, -6, 0]]))
 
     @pytest.mark.parametrize("implementation", ["scaling", "log"])
